datax-projects/requirements.txt
alabaster==0.7.9
anaconda-client==1.6.0
anaconda-navigator==1.4.3
appnope==0.1.0
appscript==1.0.1
args==0.1.0
astroid==1.4.9
astropy==1.3
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0rc1
backports.weakref==1.0rc1
beautifulsoup4==4.5.3
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.4
boto==2.45.0
Bottleneck==1.2.0
cffi==1.9.1
chardet==2.3.0
chest==0.2.3
clarifai==2.0.21
click==6.7
clint==0.5.1
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
conda==4.3.16
configobj==5.0.6
configparser==3.5.0
contextlib2==0.5.4
cryptography==1.7.1
CVXcanon==0.1.1
cvxpy==0.4.9
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.13.0
datashape==0.5.4
decorator==4.0.11
dill==0.2.5
docutils==0.13.1
ecos==2.0.4
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
flaskr==0.0.0
floyd-cli==0.9.4
future==0.15.2
gcloud==0.18.3
gevent==1.2.1
googleapis-common-protos==1.5.2
graphviz==0.6
greenlet==0.4.11
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.999
httplib2==0.10.3
idna==2.2
imageio==2.1.2
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==1.2.2
lazy-object-proxy==1.2.2
llvmlite==0.15.0
locket==0.2.0
lxml==3.7.2
MarkupSafe==0.23
marshmallow==2.13.5
matplotlib==2.0.0
mistune==0.7.3
mpld3==0.3
mpmath==0.19
multipledispatch==0.4.9
multiprocess==0.70.4
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.2
nose==1.3.7
notebook==4.3.1
numba==0.30.1
numexpr==2.6.1
numpy==1.11.3
numpydoc==0.6.0
oauth2client==4.0.0
odo==0.5.0
opencv-python==3.2.0.7
openpyxl==2.4.1
ortools-examples==5.1.4045
pandas==0.19.2
partd==0.3.7
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.9.0
ply==3.9
prompt-toolkit==1.0.9
protobuf==3.0.0
psutil==5.0.1
ptyprocess==0.5.1
PuLP==1.6.5
py==1.4.32
py3-ortools==5.1.4045
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycosat==0.6.1
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.1.3
pylint==1.6.4
pyOpenSSL==16.2.0
pyparsing==2.1.4
pytest==3.0.5
python-dateutil==2.6.0
pytube==6.2.2
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
redis==2.10.5
requests==2.13.0
requests-toolbelt==0.7.1
rope-py3k==0.9.4.post1
rsa==3.4.2
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
scs==1.2.6
seaborn==0.7.1
shortuuid==0.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.5.1
spyder==3.1.2
SQLAlchemy==1.1.5
statsmodels==0.6.1
sympy==1.0
tables==3.3.0
tabulate==0.7.7
tensorflow==1.0.0
terminado==0.6
Theano==0.8.2
toolz==0.8.2
tornado==4.4.2
traitlets==4.3.1
unicodecsv==0.14.1
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.11.15
widgetsnbextension==1.2.6
wrapt==1.10.8
xgboost==0.6
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.2
xlwt==1.2.0
